Simple Calculator using JAVA(SWITCH / IF ELSE)
It is a simple calculator in java  as name reflects it's not a rocket science but it is a major system for newbie's to coding, hope this system would help people especially new comers to coding, in this system I used both if else ladder and switch, I comment out one of them to use one method at a particular runtime. Thank You!